Spike out extraction logic
This is okay and works within its strictures, but I think
we can bring the logic forward a bit more and maybe have
a few helper utilities for common use-cases.

Thinking that the essential attributes are:

• keyCells: a list of cells representing 'primary keys'
• attributeLabelFns: a list of functions that return the
  name of the attribute (usually by looking it up in the
  header row)
• attributeValueFns: a list of functions Cell => Value, that
  most of the time will return the value of the cell's row
  in one of the columns (but might also do things like
  coerce types, etc.

After that we just iterate through the keyCells and iterate through
attributeLabelFns and attributeValueFns to generate a list of
records. We'll try that next.